dev: improve tool/dependency management
The existing dev setup had an undocumented dependency on a "global"
Python 3.10+ installation (macOS ships with 3.9), due to a quirk of
`asdf-pyapp`. The main idea of this change is to remove that plugin in
favor of using Ansible as an ordinary Python package, within a project-
local Python environment. [`uv`](https://docs.astral.sh/uv/) is used as
the package manager.

This has the nice side effects of:

* making the project's Python dependencies "visible" to Dependabot
* simplifying the CI workflow and aligning it better with local dev
* removing the need to provide configuration or separate instructions
  for ASDF alternatives like RTX and Mise

The Ansible `collections` file is also removed as unnecessary, since
this only included collections that were already bundled with `ansible`
itself. The latter has newer major versions, but the breaking changes
are minimal and don't include anything that should affect this project.