dev: improve tool/dependency management
The existing dev setup had an undocumented dependency on a "global"
Python 3.10+ installation (macOS ships with 3.9), due to a quirk of
`asdf-pyapp`. The main idea of this change is to remove that plugin in
favor of using Ansible as an ordinary Python package, within a project-
local Python environment. [`uv`](https://docs.astral.sh/uv/) is used as
a one-stop Python environment/package manager.

Some positive effects of this change:

* makes the project's Python dependencies "visible" to Dependabot,
  should we want to enable it for this repo
* simplifies the CI workflow and aligns it better with local dev
* makes it easier to run linting locally, by making the linting tools
  proper dependencies of the project
* removes the need to provide configuration or separate instructions for
  ASDF alternatives like RTX and Mise

The Ansible `collections` file is also removed as unnecessary, since
this only included collections that were already bundled with `ansible`
itself. The latter has newer major versions, but the breaking changes
are minimal and don't include anything that should affect this project.

Only minimal changes are made outside the local dev environment setup;
in particular the Ansible setup for remote hosts doesn't use `uv` or
otherwise involve the `pyproject.toml`. This might be desirable in the
future to enforce alignment of the dev and "prod" environments.